db structure update for sys user and admin modules

+++ share/db/sql/sys_admin_modules.sql.php
@@ -1,6 +1,6 @@
 <?php
 return '
-  `id` tinyint(3) unsigned NOT NULL AUTO_INCREMENT,
+  `id` int(10) unsigned NOT NULL AUTO_INCREMENT,
   `name` varchar(255) NOT NULL DEFAULT \'\',
   `description` varchar(255) NOT NULL,
   `version` varchar(16) NOT NULL,
